+++ scripts/regenerate-models.ps1
@@ -13,6 +13,12 @@ php ".\bin\phalcon-kit" cli scaffold run `
     --directory="./" `
     --src-dir="src/" `
     --namespace="App" `
+    --models-extend="App\Models\AbstractModel" `
     --force `
     --no-models `
+    --no-controllers `
+    --no-tests `
+    --no-license `
+    --protected-properties `
     $args
+exit $LASTEXITCODE